memcard: Move memory card structures into their own file
